+++ src/lib.rs
@@ -20,3 +20,4 @@ pub mod commands;
 pub mod config;
 pub mod error;
 pub mod prelude;
+pub mod cellar_wrapper;
